+++ moqt/internal/message/parameters.go
@@ -1,6 +1,0 @@
-package message
-
-/*
-* Parameters
- */
-type Parameters map[uint64][]byte